+++ Sources/StreamChat/APIClient/Endpoints/ChannelEndpoints.swift
@@ -58,14 +58,19 @@ extension Endpoint {
             body: body
         )
     }
-
-    static func muteChannel(cid: ChannelId, mute: Bool) -> Endpoint<EmptyResponse> {
-        .init(
+    
+    static func muteChannel(cid: ChannelId, mute: Bool, expiration: Int? = nil) -> Endpoint<EmptyResponse> {
+        let body: [String: AnyEncodable] = [
+            "channel_cid": AnyEncodable(cid),
+            "expiration": AnyEncodable(expiration)
+        ]
+        
+        return .init(
             path: .muteChannel(mute),
             method: .post,
             queryItems: nil,
             requiresConnectionId: true,
-            body: ["channel_cid": cid]
+            body: body
         )
     }