+++ Sources/StreamChat/Models/MuteDetails.swift
@@ -10,4 +10,6 @@ public struct MuteDetails: Equatable {
     public let createdAt: Date
     /// The time when the mute was updated.
     public let updatedAt: Date?
+    /// The expiration date of the pinning. Infinite expiration in case it is `nil`.
+    public let expiresAt: Date?
 }